Rename existing Notifier class to RequestNotifier
This gives it a more expressive name and makes it more distinct from TaskNotifier.
