+++ app/lib/index.js
@@ -1,13 +1,13 @@
 'use strict'
 
 const { BaseNotifier } = require('./base_notifier')
-const Notifier = require('./notifier')
+const RequestNotifier = require('./request_notifier')
 const StaticLookup = require('./static_lookup')
 const TaskNotifier = require('./task_notifier')
 
 module.exports = {
   BaseNotifier,
-  Notifier,
+  RequestNotifier,
   StaticLookup,
   TaskNotifier
 }
